Update index.js
Updated help to reflect lab-command change and included the leagues command.

Reworked bot status to not to change each time it joins/leaves a server to prevent API overusage if too many servers are active and to allow multiple different lines that switch over time randomly.

Added countdowns to league starts and ends if they are known.

Added support for upcoming leagues, when they are listed in the APIs